Move loader to a separate package
